+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Internal: Use [ldapts](https://github.com/ldapts/ldapts) instead of [ldapjs](https://github.com/joyent/node-ldapjs) as ldap library
+
+### Fixed
+- Fix membership resolution for ldap objects without any membership
+
+### Removed
+- *BREAKING:* LDAP StartTLS is no longer supported
+- *BREAKING:* LDAP reconnect logic
 
 ## [1.3.0] - 2019-01-07
 ### Changed